+++ test/src/customizer/widgets/customizer_body_test.mocks.dart
@@ -290,6 +290,16 @@ class MockPersistentAvatarMakerController extends _i1.Mock
       ) as _i10.Future<_i3.RestoredData>);
 
   @override
+  bool isPersistentController() => (super.noSuchMethod(
+        Invocation.method(
+          #isPersistentController,
+          [],
+        ),
+        returnValue: false,
+        returnValueForMissingStub: false,
+      ) as bool);
+
+  @override
   _i10.Future<void> initController() => (super.noSuchMethod(
         Invocation.method(
           #initController,